# fintech-bank/src/main/resources/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fintech Bank - Future of banking",
    "description": "Fun Exercise API สำหรับธนาคารการเงินยุคใหม่",
    "contact": {
      "email": "you@your-company.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
  "tags": [
    {
      "name": "account",
      "description": "Banking Account"
    }
  ],
  "paths": {
    "/accounts": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "operationId": "list",
        "responses": {
          "200": {
            "description": "list all account",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "account-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenAccount"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "create an account",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{accountNo}/withdraw": {
      "post": {
        "tags": [
          "account-controller"
        ],
        "operationId": "withdraw",
        "parameters": [
          {
            "name": "accountNo",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Withdraw"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "withdraw money from specific account",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{accountNo}/transfer/{targetAccountNo}": {
      "post": {
        "tags": [
          "account-controller"
        ],
        "operationId": "transfer",
        "parameters": [
          {
            "name": "accountNo",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "targetAccountNo",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transfer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "transfer money from specific account to specific account",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{accountNo}/deposit": {
      "post": {
        "tags": [
          "account-controller"
        ],
        "operationId": "deposit",
        "parameters": [
          {
            "name": "accountNo",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Deposit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "deposit money to specific account",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{accountNo}": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "accountNo",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "get specific account",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "OpenAccount": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "SAVING",
              "Checking",
              "CURRENT"
            ]
          },
          "name": {
            "type": "string"
          },
          "balance": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Account": {
        "type": "object",
        "properties": {
          "no": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "enum": [
              "SAVING",
              "Checking",
              "CURRENT"
            ]
          },
          "name": {
            "type": "string"
          },
          "balance": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Withdraw": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Transfer": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "remark": {
            "type": "string"
          }
        }
      },
      "Deposit": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}